move companies house module into lib
